only apply PrettyTime to Task
